+++ docs/app/[lng]/(main)/docs/page.tsx
@@ -1,14 +1,5 @@
-'use client'
-
-import { useEffect } from 'react'
 import dynamic from 'next/dynamic'
 import { Container } from '@mui/material'
-import Prism from 'prismjs'
-
-// Load the languages to highlight
-import 'prismjs/components/prism-bash'
-import 'prismjs/components/prism-jsx'
-import 'prismjs/components/prism-tsx'
 
 // Types
 import type { PageProps } from '@/types/app/pages'
@@ -17,7 +8,7 @@ import type { PageProps } from '@/types/app/pages'
 import LocaleConfig from '@/config/locale'
 
 // Components
-const IndexTemplate = dynamic(() => import('@/components/templates/docs'))
+const OverviewTemplate = dynamic(() => import('@/components/templates/docs/overview'))
 
 const HomePage = (props: PageProps) => {
 	// Props
@@ -26,14 +17,9 @@ const HomePage = (props: PageProps) => {
 	// Variables
 	const lng = params?.lng ?? LocaleConfig.default
 
-	// Callbacks
-	useEffect(() => {
-		Prism.highlightAll()
-	}, [])
-
 	return (
 		<Container maxWidth="xl">
-			<IndexTemplate lng={lng} />
+			<OverviewTemplate lng={lng} />
 		</Container>
 	)
 }