+++ docs/components/templates/docs/overview.tsx
@@ -1,0 +1,17 @@
+import dynamic from 'next/dynamic'
+
+// Types
+import type { TemplateProps } from '@/types/components/templates'
+
+// MDX
+const OverviewEn = dynamic(() => import('@/markdown/en/docs/overview.mdx'))
+const OverviewFa = dynamic(() => import('@/markdown/fa/docs/overview.mdx'))
+
+const OverviewTemplate = (props: TemplateProps) => {
+	// Props
+	const { lng } = props
+
+	return lng === 'fa' ? <OverviewFa /> : <OverviewEn />
+}
+
+export default OverviewTemplate