+++ docs/types/app/layouts/docs.d.ts
@@ -1,0 +1,10 @@
+// Types
+import type { ReactNode } from 'react'
+
+export type DocsLayoutProps = Readonly<{
+	children: ReactNode
+	params?: {
+		lng: string
+		slug: string[]
+	}
+}>