+++ docs/types/app/pages/docs.d.ts
@@ -1,0 +1,6 @@
+export type DocsPageProps = {
+	params?: {
+		lng: string
+		slug: string[]
+	}
+}