+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Added mdx files for **overview** and **installation**
+
 # 0.4.50
 - Fixed **bugs** in changing `language` and `mode` in **AppBar**